chatbot_with_chainlit/pyproject.toml
[project]
name = "hello-chainlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "raoasif", email = "raoasif452@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.4.0",
    "openai-agents>=0.0.4",
]

[project.scripts]
hello-chainlit = "hello_chainlit:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
